+++ src/components/super/i-static-page/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v??? (2023-10-??)
+
+#### :rocket: New Feature
+
+* Added the ability to build initial HTML file without any inline Javascript by configuring `webpack.externalizeInline`
+
 ## v4.0.0-beta.17 (2023-09-06)
 
 #### :bug: Bug Fix